Something gets drawn on the screen

+++ src/eliascregard/main/Settings.java
@@ -14,7 +14,7 @@ public class Settings {
     public static final double HALF_FOV = FOV / 2;
     public static final int NUM_RAYS = SCREEN_SIZE.width;
     public static final int HALF_NUM_RAYS = NUM_RAYS / 2;
-    public static final int SCALE = SCREEN_SIZE.width / NUM_RAYS;
+    public static final double SCALE = (double) SCREEN_SIZE.width / NUM_RAYS;
     public static final double DELTA_ANGLE = FOV / NUM_RAYS;
     public static final int MAX_DEPTH = 20;
     public static final Vector2 PLAYER_START_POSITION = new Vector2(1, 5);
